test(fargate): split ensure and expect into its own files
